requirements.txt
Flask==3.1.0
yt-dlp==2025.3.31
requests==2.32.3
python-dotenv==1.1.0